+++ src/test/static/javascript/module.js
@@ -1,1 +1,3 @@
 import * as submodule from './submodule.js';
+
+import * as foo from './does-not-exist.js';