+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 * [BREAKING] Removed Analyzer.createFromDirectory, the replacement will go into
   polymer-project-config.
+* [BREAKING] Import features will now be produced when a document has a broken
+  import (e.g. because the imported file does not exist, or it does not parse,
+  etc). We still emit a warning in such cases.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.0.0-pre.17] - 2018-03-20